requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dateparser==1.2.2
fastapi==0.128.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
joblib==1.5.3
keras==3.13.0
kiwisolver==1.4.9
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.0
namex==0.1.0
newsapi-python==0.2.7
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
pillow==12.1.0
propcache==0.4.1
protobuf==6.33.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-binance==1.0.34
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.1
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unicorn==2.1.4
urllib3==2.6.2
uvicorn==0.40.0
vaderSentiment==3.3.2
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
yarl==1.22.0
